requirements.txt
aiofiles==23.1.0
alembic==1.9.4
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
ecdsa==0.18.0
fastapi==0.89.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
jwt==1.3.1
Mako==1.2.4
MarkupSafe==2.1.2
mysql==0.0.3
mysqlclient==2.1.1
passlib==1.7.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
PyMySQL==1.0.2
python-git==2018.2.1
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.3
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
